Fixes encoding

+++ util.go
@@ -57,9 +57,9 @@ func write_buffer(src int32, p []byte, depth uint32) {
 func read_buffer(p []byte, depth uint32) (v int32) {
 	switch depth {
 	case 2:
-		v = int32(binary.LittleEndian.Uint16(p))
+		v = int32(int16(binary.LittleEndian.Uint16(p)))
 	case 1:
-		v = int32(p[0])
+		v = int32(int8(p[0]))
 	default:
 		v = int32(binary.LittleEndian.Uint32(p))
 	}
